requirements.txt
# Core
pynput>=1.7.6
mss>=9.0.1
Pillow>=10.0.0
customtkinter>=5.2.0
pywin32>=306
psutil>=5.9.0
jinja2>=3.1.2
python-docx>=1.1.0

# AI Providers
anthropic>=0.40.0
openai>=1.30.0

# Segra M365 Copilot integration
msal>=1.28.0

# Optional AI Providers (uncomment to install)
# google-generativeai>=0.5.0    # For Google Gemini
# Ollama uses httpx (already installed via anthropic/openai)
